Refactor out utils to init cli
+ Added extra test for query_tagging_api
