+++ samcli/commands/test_runner/init/cli.py
@@ -2,7 +2,7 @@
 CLI command for "test-runner init" command
 """
 import logging
-from typing import Any, List, Optional
+from typing import Any, List, Optional, Union
 
 import click
 
@@ -13,8 +13,7 @@ from samcli.commands.exceptions import NoResourcesMatchGivenTagException
 from samcli.lib.test_runner.generate_env_vars import FargateRunnerArnMapGenerator
 from samcli.lib.test_runner.invoke_testsuite import FargateTestsuiteRunner
 from samcli.lib.test_runner.test_runner_template_generator import FargateRunnerCFNTemplateGenerator
-from samcli.lib.test_runner.utils import query_tagging_api, write_file
-from samcli.lib.utils.boto_utils import get_boto_client_provider_with_config
+from samcli.lib.utils.boto_utils import BotoProviderType, get_boto_client_provider_with_config
 from samcli.lib.utils.colors import Colored
 
 SHORT_HELP = "Generates a Test Runner CloudFormation Template in YAML format, as well as a resource-ARN map."
@@ -145,7 +144,7 @@ def do_cli(
 def _create_arn_map_file(env_file: str, resource_arn_list: List[str]) -> None:
     arn_map_generator = FargateRunnerArnMapGenerator()
     resource_arn_map_yaml_string = arn_map_generator.generate_env_vars_yaml_string(resource_arn_list)
-    write_file(env_file, resource_arn_map_yaml_string)
+    _write_file(env_file, resource_arn_map_yaml_string)
 
     LOG.info(
         COLOR.green(f"✓ Sucessfully generated an environment variable specification file `test_runner_arn_map.yaml`.\n")
@@ -162,7 +161,7 @@ def _create_arn_map_file(env_file: str, resource_arn_list: List[str]) -> None:
 def _create_test_runner_template(image_uri: str, template_name: str, resource_arn_list: List[str]) -> None:
     template_generator = FargateRunnerCFNTemplateGenerator(resource_arn_list)
     test_runner_cfn_contents = template_generator.generate_test_runner_template_string(image_uri)
-    write_file(template_name, test_runner_cfn_contents)
+    _write_file(template_name, test_runner_cfn_contents)
     LOG.info(
         COLOR.green(f"\n✓ Successfully generated a Test Runner CloudFormation Template named `{template_name}`!\n")
     )
@@ -176,3 +175,57 @@ def _create_test_runner_template(image_uri: str, template_name: str, resource_ar
                 "Make any other changes you wish, and when you're ready to run your tests, use `sam test-runner run.`\n"
             )
         )
+
+
+def _get_image_uri(runtime: str) -> str:
+    # TODO: Public ECR repositories not yet set up.
+    #       When ready, public test-running image URIs will be returned corresponding to the specified runtime
+    pass
+
+
+def _write_file(filename: str, contents: str) -> None:
+    with open(filename, "w") as f:
+        f.write(contents)
+
+
+def query_tagging_api(tags: dict, boto_client_provider: BotoProviderType) -> Union[List[str], None]:
+    """
+    Queries the Tagging API to retrieve the ARNs of every resource with the given tags.
+
+    NOTE: https://docs.aws.amazon.com/cli/latest/reference/resourcegroupstaggingapi/get-resources.html#get-resources
+
+    Parameters
+    ----------
+    tag_filters : dict
+        The tag filters to restrict output to only those resources with the given tags.
+
+        Takes the form `{'Key1':'Value1', 'Key2':'Value2', ...}`
+
+
+    boto_client_provider : BotoProviderType
+        Provides a boto3 client in order to query tagging API
+
+    Returns
+    -------
+    dict
+        A ResourceTagMappingList, which contains a list of resource ARNs and tags associated with each.
+
+        NOTE: https://docs.aws.amazon.com/cli/latest/reference/resourcegroupstaggingapi/get-resources.html#output
+
+    Raises
+    ------
+    botocore.ClientError
+        If the `get_resources` call fails.
+
+        NOTE: # https://boto3.amazonaws.com/v1/documentation/api/latest/guide/error-handling.html#parsing-error-responses-and-catching-exceptions-from-aws-services
+    """
+    # Convert tag format into one that the API accepts
+    # NOTE: https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/resourcegroupstaggingapi.html#ResourceGroupsTaggingAPI.Client.get_resources
+    tag_filters = [{"Key": k, "Values": [v]} for k, v in tags.items()]
+    resource_tag_mapping_list = (
+        boto_client_provider("resourcegroupstaggingapi")
+        .get_resources(TagFilters=tag_filters)
+        .get("ResourceTagMappingList")
+    )
+
+    return [resource.get("ResourceARN") for resource in resource_tag_mapping_list]
